requirements.txt
asgiref==3.3.4
Django==3.2
django-phonenumber-field==5.0.0
phonenumberslite==8.12.21
pytz==2021.1
sqlparse==0.4.1
